requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
requests==2.26.0
python-dotenv==0.19.0
pytest==6.2.5
httpx==0.19.0
python-multipart==0.0.5